+++ __test__/src/lib.rs
@@ -3,12 +3,13 @@ mod tests {
     use std::{io, time::SystemTime};
 
     use rowid::{
-        decode, encode, generate, get_randomness, rowid,
-        system_time_to_timestamp, timestamp_to_system_time, verify,
-        GenerateResult, RowIDWithConfig, RowIDWithConfigResult, VerifyResult,
+        base::{
+            decode, encode, generate, get_randomness, rowid, verify,
+            GenerateResult, RowIDError, VerifyResult,
+        },
+        time::{system_time_to_timestamp, timestamp_to_system_time},
+        with_config::{RowIDWithConfig, RowIDWithConfigResult},
     };
-
-    use rowid::RowIDError;
 
     // system_time_to_timestamp