+++ package/src/base/mod.rs
@@ -1,2 +1,0 @@
-pub mod rowid;
-pub mod rowid_with_config;